Backend/auth/requirements.txt
# Twilio Authentication Dependencies
twilio>=8.10.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
